Refactored appointment-booking after locations, sign-in, and datetime.
Shared BookingDetailCallout and SearchRow, collapsed ServiceLocation into one type, bumped React Router to 8.3 and CI/Docker-aligned Node to 24, and replaced license-checker with license-checker-rseidelsohn.
