+++ appointment-booking/package.json
@@ -11,7 +11,7 @@
     "format": "prettier --write app",
     "format:check": "prettier --check app",
     "type-check": "react-router typegen && tsc",
-    "license-check": "license-checker --production --onlyAllow \"MIT;Apache-2.0;ISC;BSD-2-Clause;BSD-3-Clause;0BSD;Unlicense\" --excludePackages \"appointment-booking@0.0.0;@bcgov/bc-sans@2.1.0\"",
+    "license-check": "license-checker-rseidelsohn --production --onlyAllow \"MIT;Apache-2.0;ISC;BSD-2-Clause;BSD-3-Clause;0BSD;Unlicense\" --excludePackages \"appointment-booking@0.0.0;@bcgov/bc-sans@2.1.0\"",
     "ci:check": "npm run lint && npm run type-check && npm run build && npm run license-check"
   },
   "dependencies": {
@@ -22,18 +22,18 @@
     "@fortawesome/free-solid-svg-icons": "^7.3.0",
     "@fortawesome/react-fontawesome": "^3.3.1",
     "@internationalized/date": "^3.12.2",
-    "@react-router/node": "8.0.0",
-    "@react-router/serve": "8.0.0",
+    "@react-router/node": "8.3.0",
+    "@react-router/serve": "8.3.0",
     "geolib": "^3.3.14",
     "isbot": "^5.1.36",
     "keycloak-js": "^26.2.4",
     "react": "^19.2.6",
     "react-dom": "^19.2.6",
-    "react-router": "8.0.0"
+    "react-router": "8.3.0"
   },
   "devDependencies": {
     "@eslint/js": "^10.0.1",
-    "@react-router/dev": "8.0.0",
+    "@react-router/dev": "8.3.0",
     "@types/node": "^24.12.3",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
@@ -45,7 +45,7 @@
     "eslint-plugin-react-hooks": "^7.1.1",
     "eslint-plugin-react-refresh": "^0.5.2",
     "globals": "^17.6.0",
-    "license-checker": "^25.0.1",
+    "license-checker-rseidelsohn": "^5.0.1",
     "prettier": "^3.8.3",
     "typescript": "~6.0.2",
     "typescript-eslint": "^8.59.2",
