+++ appointment-booking/vite.config.ts
@@ -1,12 +1,15 @@
 import { reactRouter } from '@react-router/dev/vite'
 import path from 'path'
+import { fileURLToPath } from 'url'
 import { defineConfig } from 'vite'
+
+const rootDir = path.dirname(fileURLToPath(import.meta.url))
 
 export default defineConfig({
   plugins: [reactRouter()],
   resolve: {
     alias: {
-      '~': path.resolve(__dirname, './app'),
+      '~': path.resolve(rootDir, './app'),
     },
   },
 })
